sources: refactor subversion source into module.
Extract the Subversion source into `sources/_subversion.py`.

LP: #1648906
