+++ snapcraft/internal/sources/__init__.py
@@ -73,7 +73,6 @@ import os
 import os.path
 import re
 import shutil
-import subprocess
 import tempfile
 import tarfile
 import zipfile
@@ -81,13 +80,14 @@ import zipfile
 from snapcraft.internal import common
 from . import errors
 from . import _base
-from ._bazaar import Bazaar        # noqa
-from ._deb import Deb              # noqa
-from ._git import Git              # noqa
-from ._local import Local          # noqa
-from ._mercurial import Mercurial  # noqa
-from ._rpm import Rpm              # noqa
-from ._script import Script        # noqa
+from ._bazaar import Bazaar          # noqa
+from ._deb import Deb                # noqa
+from ._git import Git                # noqa
+from ._local import Local            # noqa
+from ._mercurial import Mercurial    # noqa
+from ._rpm import Rpm                # noqa
+from ._script import Script          # noqa
+from ._subversion import Subversion  # noqa
 
 logging.getLogger('urllib3').setLevel(logging.CRITICAL)
 
@@ -105,48 +105,6 @@ __SOURCE_DEFAULTS = {
 
 def get_source_defaults():
     return __SOURCE_DEFAULTS.copy()
-
-
-class Subversion(_base.Base):
-
-    def __init__(self, source, source_dir, source_tag=None, source_commit=None,
-                 source_branch=None, source_depth=None):
-        super().__init__(source, source_dir, source_tag, source_commit,
-                         source_branch, source_depth, 'svn')
-        if source_tag:
-            if source_branch:
-                raise errors.IncompatibleOptionsError(
-                    "Can't specify source-tag OR source-branch for a "
-                    "Subversion source")
-            else:
-                raise errors.IncompatibleOptionsError(
-                    "Can't specify source-tag for a Subversion source")
-        elif source_branch:
-            raise errors.IncompatibleOptionsError(
-                "Can't specify source-branch for a Subversion source")
-        if source_depth:
-            raise errors.IncompatibleOptionsError(
-                'can\'t specify source-depth for a Subversion source')
-
-    def pull(self):
-        opts = []
-
-        if self.source_commit:
-            opts = ["-r", self.source_commit]
-
-        if os.path.exists(os.path.join(self.source_dir, '.svn')):
-            subprocess.check_call(
-                [self.command, 'update'] + opts, cwd=self.source_dir)
-        else:
-            if os.path.isdir(self.source):
-                subprocess.check_call(
-                    [self.command, 'checkout',
-                     'file://{}'.format(os.path.abspath(self.source)),
-                     self.source_dir] + opts)
-            else:
-                subprocess.check_call(
-                    [self.command, 'checkout', self.source, self.source_dir] +
-                    opts)
 
 
 class Tar(_base.FileBase):
